Create commands package and move handlers
