+++ tests/test_commands.py
@@ -7,7 +7,7 @@ import pytest
 import sys
 sys.path.append(str(Path(__file__).resolve().parent.parent))
 
-import commands
+from commands import core as commands
 
 
 @pytest.fixture
